+++ libs/artifact/CHANGELOG.rst
@@ -1,3 +1,7 @@
+**1.1.0 - 08/25/26**
+
+- **Breaking change.** Correct ``requires-python`` to ``>=3.11`` to match the tested versions
+
 **1.0.9 - 07/31/26**
 
 - Make the ``hdf`` module private