+++ libs/engine/CHANGELOG.rst
@@ -1,3 +1,12 @@
+**5.7.0 - 08/25/26**
+
+- **Breaking change.** Drop support for Python 3.10 and add support for Python 3.13
+
+**5.6.1 - 08/24/26**
+
+- Declare ``InteractiveContext``'s full constructor signature and default ``logging_verbosity`` to 0
+- Document the ``SimulationContext`` and ``InteractiveContext`` constructors
+
 **5.6.0 - 08/06/26**
 
 - Support pandas 3